pyproject.toml
[project]
name = "bio_aho_tagger"
version = "0.1.0"
description = "Baseline dictionary based package to resolve chemical structures and tag biological entities."
readme = "README.md"
requires-python = ">=3.10"
dependencies = ["pyahocorasick==2.1.0"]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build]
include = ["bio_aho_tagger/data/chembl_smiles.pkl", "bio_aho_tagger/data/EFO_disease.pkl"]

[tool.hatch.build.targets.wheel]
packages = ["src/bio_aho_tagger"]

[project.scripts]
create-tagger-dict = "bio_aho_tagger.scripts:create_file"